moved App.jsx file to fix build error and udpated imports
