+++ src/Components/ProjectsPage.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useContext, useState } from "react";
 import { Container, Button, Modal } from "react-bootstrap";
 import AOS from "aos";
-import { ThemeContext } from "./App";
+import { ThemeContext } from "../App";
 import GitHubSVG from "./svg/GitHubSVG";
 import VideoSVG from "./svg/VideoSVG";
 import OpenLinkSVG from "./svg/OpenLinkSVG";
